# maas-service/docs/swagger.yaml
definitions:
...
  model.Account:
    properties:
      domainNamespaces:
        example:
        - namespace1
        items:
          type: string
        type: array
      manager:
        example:
        - client
        items:
          type: string
        type: array
      namespace:
        example: _GLOBAL
        type: string
      password:
        example: '***'
        type: string
      salt:
        example: '***'
        type: string
      username:
        example: manager
        type: string
    type: object
  model.BgStatus:
    properties:
      active:
        example: v2
        type: string
      candidates:
        example:
        - v2
        - v3
        items:
          type: string
        type: array
      legacy:
        example: v3
        type: string
      namespace:
        example: namespace
        type: string
      timestamp:
        example: "2021-08-18T17:29:06.82837Z"
        type: string
    type: object
...
  model.ClientAccountDto:
    properties:
      namespace:
        example: namespace
        type: string
      password:
        example: client
        type: string
      roles:
        example:
        - agent
        items:
          type: string
        type: array
      username:
        example: client
        type: string
    type: object
  model.CpDeploymentVersion:
    properties:
      createdWhen:
        example: "2021-08-18T16:33:15.142354459Z"
        type: string
      stage:
        example: ACTIVE
        type: string
      updatedWhen:
        example: "2021-08-18T16:33:15.14235456Z"
        type: string
      version:
        example: v1
        type: string
    type: object
...
  model.ManagerAccountDto:
    properties:
      password:
        example: 24002eadbc
        type: string
      username:
        example: manager
        type: string
    type: object
...
  v1.RegistrationRequest:
    properties:
      id:
        type: string
      namespaces:
        items:
          type: string
        type: array
    required:
    - id
    - namespaces
    type: object
  v1.RegistrationResponse:
    properties:
      id:
        type: string
      namespaces:
        items:
          type: string
        type: array
    type: object
info:
  contact: {}
  description: API for Maas Service.
  title: Maas Service API
paths:
  /api/composite/v1/structure:
    post:
      consumes:
      - application/json
      description: Register new Composite
      operationId: Create
      parameters:
      - description: RegistrationRequest
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.RegistrationRequest'
      produces:
      - application/json
      responses:
        "204":
          description: ""
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BasicAuth:
        - agent
      summary: Create Composite Registration
      tags:
      - V1
  /api/composite/v1/structure/{id}:
    delete:
      description: Destroy Composite Registrations
      operationId: DeleteById
      parameters:
      - description: Registration ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: ""
        "404":
          description: ""
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BasicAuth:
        - agent
      summary: Destroy Composite Registrations
      tags:
      - V1
    get:
      description: Get Composite Registration By Baseline ID
      operationId: GetById
      parameters:
      - description: Registration ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.RegistrationResponse'
        "404":
          description: ""
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BasicAuth:
        - agent
      summary: Get Composite Registration By ID
      tags:
      - V1
  /api/composite/v1/structures:
    get:
      description: Get All Composite Registrations
      operationId: GetAll
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/v1.RegistrationResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BasicAuth:
        - agent
      summary: Get All Composite Registrations
      tags:
      - V1
  /api/v1/auth/account/client:
    delete:
      description: Delete Client Account
      operationId: DeleteClientAccount
      parameters:
      - description: ClientAccountDto
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.ClientAccountDto'
      produces:
      - application/json
      responses:
        "204":
          description: ""
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BasicAuth:
        - manager
      summary: Delete Client Account using Manager Role
      tags:
      - V1
    post:
      description: Create new account
      operationId: CreateAccount
      parameters:
      - description: ClientAccountDto
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.ClientAccountDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.ClientAccountDto'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BasicAuth:
        - manager
      summary: Create Account
      tags:
      - V1
  /api/v1/auth/account/manager:
    post:
      description: Save Manager Account
      operationId: SaveManagerAccount
      parameters:
      - description: ManagerAccountDto
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.ManagerAccountDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.ManagerAccountDto'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BasicAuth:
        - manager
      summary: Save Manager Account using Manager Role
      tags:
      - V1
  /api/v1/auth/accounts:
    get:
      description: Get All Accounts
      operationId: GetAllAccounts
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Account'
            type: array
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BasicAuth:
        - manager
      summary: Get All Accounts using Manager Role
      tags:
      - V1
  /api/v1/bg-status:
    get:
      description: Get Bg Status By Namespace
      operationId: GetBgStatusByNamespace
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.BgStatus'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BasicAuth:
        - agent
      summary: Get Bg Status By Namespace using Agent Role
      tags:
      - V1
    post:
      description: Apply Bg Status
      operationId: ApplyBgStatus
      parameters:
      - description: CpMessageDto
        in: body
        name: request
        required: true
        schema:
          items:
            $ref: '#/definitions/model.CpDeploymentVersion'
          type: array
      - description: X-Origin-Namespace
        in: header
        name: X-Origin-Namespace
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BasicAuth:
        - agent
      summary: Apply Bg Status using Agent Role
      tags:
      - V1
...
  /api/v2/auth/account/manager/{name}/password:
    put:
      description: Update Password
      operationId: UpdatePassword
      parameters:
      - description: new-password
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BasicAuth:
        - manager
      summary: Update Password using Manager Role
      tags:
      - V2
...
produces:
- application/json
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"